// src/Mobile/ReportsApp/services/api/models/CheckListQueryResult.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Inspections API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    CheckListItemQueryResult,
    CheckListItemQueryResultFromJSON,
    CheckListItemQueryResultFromJSONTyped,
    CheckListItemQueryResultToJSON,
} from './';

/**
 * 
 * @export
 * @interface CheckListQueryResult
 */
export interface CheckListQueryResult {
    /**
     * 
     * @type {number}
     * @memberof CheckListQueryResult
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof CheckListQueryResult
     */
    reportId?: number;
    /**
     * 
     * @type {string}
     * @memberof CheckListQueryResult
     */
    text?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CheckListQueryResult
     */
    annotation?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof CheckListQueryResult
     */
    checked?: boolean;
    /**
     * 
     * @type {Array<CheckListItemQueryResult>}
     * @memberof CheckListQueryResult
     */
    checks?: Array<CheckListItemQueryResult> | null;
}

export function CheckListQueryResultFromJSON(json: any): CheckListQueryResult {
    return CheckListQueryResultFromJSONTyped(json, false);
}

export function CheckListQueryResultFromJSONTyped(json: any, ignoreDiscriminator: boolean): CheckListQueryResult {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'reportId': !exists(json, 'reportId') ? undefined : json['reportId'],
        'text': !exists(json, 'text') ? undefined : json['text'],
        'annotation': !exists(json, 'annotation') ? undefined : json['annotation'],
        'checked': !exists(json, 'checked') ? undefined : json['checked'],
        'checks': !exists(json, 'checks') ? undefined : (json['checks'] === null ? null : (json['checks'] as Array<any>).map(CheckListItemQueryResultFromJSON)),
    };
}

export function CheckListQueryResultToJSON(value?: CheckListQueryResult | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'reportId': value.reportId,
        'text': value.text,
        'annotation': value.annotation,
        'checked': value.checked,
        'checks': value.checks === undefined ? undefined : (value.checks === null ? null : (value.checks as Array<any>).map(CheckListItemQueryResultToJSON)),
    };
}