// src/Mobile/ReportsApp/services/api/models/CheckListItemQueryResult.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Inspections API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CheckListItemQueryResult
 */
export interface CheckListItemQueryResult {
    /**
     * 
     * @type {number}
     * @memberof CheckListItemQueryResult
     */
    checkListId?: number;
    /**
     * 
     * @type {number}
     * @memberof CheckListItemQueryResult
     */
    checked?: number;
    /**
     * 
     * @type {boolean}
     * @memberof CheckListItemQueryResult
     */
    editable?: boolean;
    /**
     * 
     * @type {number}
     * @memberof CheckListItemQueryResult
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof CheckListItemQueryResult
     */
    remarks?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof CheckListItemQueryResult
     */
    required?: boolean;
    /**
     * 
     * @type {string}
     * @memberof CheckListItemQueryResult
     */
    text?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof CheckListItemQueryResult
     */
    touched?: boolean;
}

export function CheckListItemQueryResultFromJSON(json: any): CheckListItemQueryResult {
    return CheckListItemQueryResultFromJSONTyped(json, false);
}

export function CheckListItemQueryResultFromJSONTyped(json: any, ignoreDiscriminator: boolean): CheckListItemQueryResult {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'checkListId': !exists(json, 'checkListId') ? undefined : json['checkListId'],
        'checked': !exists(json, 'checked') ? undefined : json['checked'],
        'editable': !exists(json, 'editable') ? undefined : json['editable'],
        'id': !exists(json, 'id') ? undefined : json['id'],
        'remarks': !exists(json, 'remarks') ? undefined : json['remarks'],
        'required': !exists(json, 'required') ? undefined : json['required'],
        'text': !exists(json, 'text') ? undefined : json['text'],
        'touched': !exists(json, 'touched') ? undefined : json['touched'],
    };
}

export function CheckListItemQueryResultToJSON(value?: CheckListItemQueryResult | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'checkListId': value.checkListId,
        'checked': value.checked,
        'editable': value.editable,
        'id': value.id,
        'remarks': value.remarks,
        'required': value.required,
        'text': value.text,
        'touched': value.touched,
    };
}